fix movement while conversation bug
